Export separate CMake targets for shared and for static libraries.
Allow reverse dependencies that use CMake to link to the static *and* to
the shared libaec and/or SZIP libraries in the same project.
Do that in a backward-compatible way by providing additional CMake import
targets `libaec::aec-shared`, `libaec::sz-shared`, `libaec::aec-static`,
and `libaec::sz-static` and by continuing to provide the CMake import
targets `libaec::aec` and `libaec::sz`. The latter two continue to
correspond to the shared or static libraries depending on the value that
a user set for `libaec_USE_STATIC_LIBS`.
The same holds for the value of the variables `libaec_LIBRARY` and
`SZIP_LIBRARY`. Additionally, the variables `libaec-shared_LIBRARY`,
`libaec-static_LIBRARY`, `SZIP-shared_LIBRARY`, and `SZIP-static_LIBRARY`
are set by the CMake config file.