Optionally, build only shared or only static libraries.
Add configuration options `BUILD_SHARED_LIBS` and `BUILD_STATIC_LIBS`
that can be used to select whether the shared and/or the static libraries
should be build. The names for these options are widely used for this
purpose by many projects that use CMake as their build system.

Additionally, set the `PUBLIC_HEADERS` property for both the static and
the shared libraries.

This change covers parts of the changes by @bansan85 in
https://github.com/bansan85/libaec/commit/00203c87449e749d763026b5065365e81a9dedb4.
But it tries to use more "standardized" names for the configure options.